backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djoser>=2.1.0
django-filter>=23.1
reportlab>=4.0.4
Pillow==10.3.0
gunicorn==22.0.0
psycopg2-binary>=2.9.9
python-dotenv>=0.21.0
drf-extra-fields
django-extensions